+++ src/util/hit_result.rs
@@ -3,9 +3,9 @@ use crate::util::vec3::Vec3;
 #[derive(Copy, Clone)]
 pub struct HitResult {
     pub normal: Vec3,
-    pub t: f64,
+    pub t: f32,
     pub point: Vec3,
     pub material_index: usize,
-    pub u: f64,
-    pub v: f64,
+    pub u: f32,
+    pub v: f32,
 }
